+++ packages/astro/test/fixtures/alias-tsconfig/src/ts-file.ts
@@ -1,0 +1,3 @@
+export function whoImI() {
+	return "I'm a TypeScript file!";
+}